Fixed a lot

+++ Hermes/Modules/Emojis/Emdelete.cs
@@ -1,6 +1,8 @@
-using System.Threading.Tasks;
 using Discord;
 using Hermes.Modules.Services;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Hermes.Modules.Emojis
 {
@@ -8,7 +10,7 @@ namespace Hermes.Modules.Emojis
     public class Emdelete : CommandModuleBase
     {
         [RequiredUserPermissions(GuildPermission.ManageEmojis)]
-        [DiscordCommand("emdelete", description ="Deletes given emoji.", example ="emdelete kekw", commandHelp ="emrename emoji_name")]
+        [DiscordCommand("emdelete", description = "Deletes given emoji.", example = "emdelete kekw", commandHelp = "emrename emoji_name")]
         public async Task EMDEL(params string[] args)
         {
             if (args.Length == 0)
@@ -32,13 +34,41 @@ namespace Hermes.Modules.Emojis
                 return;
             }
             var i = await GetEmote(args[0]);
-            await Context.Guild.DeleteEmoteAsync(i);
-            await ReplyAsync(embed: new EmbedBuilder
+            Emote cros = Emote.Parse("<a:cros:859033035545378826>");
+            Emote tickk = Emote.Parse("<a:tick:859032462410907649>");
+            var gc = Guid.NewGuid();
+            var cb = new ComponentBuilder().
+                WithButton("", $"{gc}Tick", ButtonStyle.Secondary, tickk).
+                WithButton("", $"{gc}Cros", ButtonStyle.Secondary, cros);
+            await Context.Channel.SendMessageAsync($"Are you sure you want to delete {i}?\nThis is a potentially destructive action.", component: cb.Build());
+            CancellationTokenSource cancelSource = new CancellationTokenSource();
+            cancelSource.CancelAfter(15000);
+            var Interaction = await InteractionHandler.NextButtonAsync(k => k.Data.CustomId.Contains(gc.ToString()) && k.User.Id == Context.User.Id, cancelSource.Token);
+            if (Interaction == null)
             {
-                Title = "Emoji Deleted Successfully!",
-                Description = "The emoji was deleted",
+                await ReplyAsync("No response received!");
+                return;
+            }
+            var isTick = Interaction.Data.CustomId.Contains("Tick");
+            if (!isTick)
+            {
+                await Context.Channel.SendMessageAsync("", false, new EmbedBuilder
+                {
+                    Title = "Alright then...",
+                    Color = Blurple,
+                    ImageUrl = "https://i.imgur.com/RBC7KUt.png"
+                }.WithCurrentTimestamp().Build());
+                return;
+            }
+            var in_ = i.Name;
+            await Context.Guild.DeleteEmoteAsync(i);
+            await ReplyAsync("", false, new EmbedBuilder
+            {
+                Title = "Action successful!",
+                Description = $"`:{in_}:` was deleted successfully!",
                 Color = Blurple
             }.WithCurrentTimestamp());
+
         }
     }
 }
